Converted add item page to view convention
